Add files via upload
Insurance Fraud Detection :

Fraud is causing billions of $$ in loss for insurance industry. This project has attempted to develop a ML algorithm to detect. The project has used the historical transaction data including normal transactions and fraud ones to obtain normal/fraud behavior features based on machine learning techniques, and utilized these features to check if a transaction is fraud or not. A cmparative study has been conducted to decide which classifier is best for this project to train the behavior features of normal and abnormal transactions.

Objective :

The objective is to construct a model to predict which transactions could be fraudulent with high accuracy.

Plan :

The data that I have is from Automobile Insurance. I will be creating a predictive model that predicts if an insurance claim is fraudulent or not. The answere between YES/NO, is a Binary Classification task. This report deals with classification algorithm random forest model to detect fraud transaction in Python.